start to tidy-up test_features.py
