+++ tests/bofire/data_models/test_features.py
@@ -15,7 +15,6 @@ from bofire.data_models.features.api import (
     CategoricalInput,
     CategoricalMolecularInput,
     CategoricalOutput,
-    ContinuousDescriptorInput,
     ContinuousInput,
     ContinuousOutput,
     DiscreteInput,
@@ -35,1220 +34,6 @@ from bofire.data_models.surrogates.api import ScalerEnum
 objective = MinimizeObjective(w=1)
 
 RDKIT_AVAILABLE = importlib.util.find_spec("rdkit") is not None
-
-
-@pytest.mark.parametrize(
-    "spec, n",
-    [
-        (spec, n)
-        for spec in specs.features.valids
-        if (spec.cls != ContinuousOutput)
-        and (spec.cls != MolecularInput)
-        and (spec.cls != CategoricalOutput)
-        for n in [1, 5]
-    ],
-)
-def test_sample(spec: specs.Spec, n: int):
-    feat = spec.obj()
-    samples = feat.sample(n=n)
-    feat.validate_candidental(samples)
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected, expected_value",
-    [
-        (ContinuousInput(key="k", bounds=(1, 1)), True, [1]),
-        (ContinuousInput(key="k", bounds=(1, 2)), False, None),
-        (ContinuousInput(key="k", bounds=(2, 3)), False, None),
-        (
-            ContinuousDescriptorInput(
-                key="k",
-                bounds=(1, 1),
-                descriptors=["a", "b"],
-                values=[1, 2],
-            ),
-            True,
-            [1],
-        ),
-        (
-            ContinuousDescriptorInput(
-                key="k",
-                bounds=(1, 2),
-                descriptors=["a", "b"],
-                values=[1, 2],
-            ),
-            False,
-            None,
-        ),
-        (
-            ContinuousDescriptorInput(
-                key="k",
-                bounds=(2, 3),
-                descriptors=["a", "b"],
-                values=[1, 2],
-            ),
-            False,
-            None,
-        ),
-    ],
-)
-def test_continuous_input_feature_is_fixed(input_feature, expected, expected_value):
-    assert input_feature.is_fixed() == expected
-    assert input_feature.fixed_value() == expected_value
-
-
-def test_continuous_input_invalid_stepsize():
-    with pytest.raises(ValueError):
-        ContinuousInput(key="a", bounds=(1, 1), stepsize=0)
-    with pytest.raises(ValueError):
-        ContinuousInput(key="a", bounds=(0, 5), stepsize=0.3)
-    with pytest.raises(ValueError):
-        ContinuousInput(key="a", bounds=(0, 1), stepsize=1)
-
-
-def test_continuous_input_round():
-    feature = ContinuousInput(key="a", bounds=(0, 5))
-    values = pd.Series([1.0, 1.3, 0.55])
-    assert_series_equal(values, feature.round(values))
-    feature = ContinuousInput(key="a", bounds=(0, 5), stepsize=0.25)
-    assert_series_equal(pd.Series([1.0, 1.25, 0.5]), feature.round(values))
-    feature = ContinuousInput(key="a", bounds=(0, 5), stepsize=0.1)
-    assert_series_equal(pd.Series([1.0, 1.3, 0.5]), feature.round(values))
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected",
-    [
-        (
-            ContinuousInput(key="if1", bounds=(0.5, 4)),
-            (0.5, 4.0),
-        ),
-        (ContinuousInput(key="if1", bounds=(2.5, 2.9)), (1, 3.0)),
-        (ContinuousInput(key="if2", bounds=(1, 3)), (1, 3.0)),
-        (ContinuousInput(key="if2", bounds=(1, 1)), (1, 1.0)),
-    ],
-)
-def test_continuous_input_feature_get_bounds(input_feature, expected):
-    experiments = pd.DataFrame({"if1": [1.0, 2.0, 3.0], "if2": [1.0, 1.0, 1.0]})
-    lower, upper = input_feature.get_bounds(values=experiments[input_feature.key])
-    assert (lower[0], upper[0]) == expected
-    lower, upper = input_feature.get_bounds()
-    assert (lower[0], upper[0]) == (
-        input_feature.lower_bound,
-        input_feature.upper_bound,
-    )
-
-
-@pytest.mark.parametrize(
-    "input_feature, values, strict",
-    [
-        (
-            specs.features.valid(ContinuousInput).obj(),
-            pd.Series([random.uniform(3.0, 5.3) for _ in range(20)]),
-            True,
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(),
-            pd.Series([random.uniform(3.0, 5.3) for _ in range(20)]),
-            False,
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(bounds=(3, 3)),
-            pd.Series([random.uniform(3.0, 5.3) for _ in range(20)]),
-            True,
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(bounds=(3, 3)),
-            pd.Series([random.uniform(3.0, 5.3) for _ in range(20)]),
-            False,
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(bounds=(3, 3)),
-            pd.Series([3.0, 3.0, 3.0]),
-            False,
-        ),
-    ],
-)
-def test_continuous_input_feature_validate_valid(input_feature, values, strict):
-    input_feature.validate_experimental(values, strict)
-
-
-@pytest.mark.parametrize(
-    "input_feature, values, strict",
-    [
-        (
-            specs.features.valid(ContinuousInput).obj(),
-            pd.Series([3.0, "mama"]),
-            True,
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(),
-            pd.Series([3.0, "mama"]),
-            False,
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(bounds=(3, 3)),
-            pd.Series([3.0, 3.0, 3.0]),
-            True,
-        ),
-    ],
-)
-def test_continuous_input_feature_validate_invalid(input_feature, values, strict):
-    with pytest.raises(ValueError):
-        input_feature.validate_experimental(values, strict)
-
-
-@pytest.mark.parametrize(
-    "input_feature, values",
-    [
-        (
-            specs.features.valid(ContinuousInput).obj(),
-            pd.Series([random.uniform(3.0, 5.3) for _ in range(20)]),
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(bounds=(3, 3)),
-            pd.Series([3.0, 3.0, 3.0]),
-        ),
-    ],
-)
-def test_continuous_input_feature_validate_candidental_valid(input_feature, values):
-    input_feature.validate_candidental(values)
-
-
-@pytest.mark.parametrize(
-    "input_feature, values",
-    [
-        (
-            specs.features.valid(ContinuousInput).obj(),
-            pd.Series([3.1, "a"]),
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(),
-            pd.Series([2.9, 4.0]),
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(),
-            pd.Series([4.0, 6]),
-        ),
-        (
-            specs.features.valid(ContinuousInput).obj(bounds=(3, 3)),
-            pd.Series([3.1, 3.2, 3.4]),
-        ),
-    ],
-)
-def test_continuous_input_feature_validate_candidental_invalid(input_feature, values):
-    with pytest.raises(ValueError):
-        input_feature.validate_candidental(values)
-
-
-@pytest.mark.parametrize(
-    "feature, xt, expected",
-    [
-        (
-            ContinuousInput(key="a", bounds=(0, 10)),
-            pd.Series(np.linspace(0, 1, 11)),
-            np.linspace(0, 10, 11),
-        ),
-        (
-            ContinuousInput(key="a", bounds=(-10, 20)),
-            pd.Series(np.linspace(0, 1)),
-            np.linspace(-10, 20),
-        ),
-    ],
-)
-def test_continuous_input_feature_from_unit_range(feature, xt, expected):
-    x = feature.from_unit_range(xt)
-    assert np.allclose(x.values, expected)
-
-
-@pytest.mark.parametrize(
-    "feature, x, expected, real",
-    [
-        (
-            ContinuousInput(key="a", bounds=(0, 10)),
-            pd.Series(np.linspace(0, 10, 11)),
-            np.linspace(0, 1, 11),
-            True,
-        ),
-        (
-            ContinuousInput(key="a", bounds=(-10, 20)),
-            pd.Series(np.linspace(-10, 20)),
-            np.linspace(0, 1),
-            True,
-        ),
-        (
-            ContinuousInput(key="a", bounds=(0, 10)),
-            pd.Series(np.linspace(0, 10, 11)),
-            np.linspace(0, 1, 11),
-            False,
-        ),
-        (
-            ContinuousInput(key="a", bounds=(-10, 20)),
-            pd.Series(np.linspace(-10, 20)),
-            np.linspace(0, 1),
-            False,
-        ),
-        (
-            ContinuousInput(key="a", bounds=(0, 9)),
-            pd.Series(np.linspace(0, 10, 11)),
-            np.linspace(0, 1, 11),
-            True,
-        ),
-        (
-            ContinuousInput(key="a", bounds=(0, 9)),
-            pd.Series(np.linspace(0, 10, 11)),
-            np.linspace(0, 10 / 9, 11),
-            False,
-        ),
-    ],
-)
-def test_continuous_input_feature_to_unit_range(feature, x, expected, real):
-    xt = feature.to_unit_range(x)
-    assert np.allclose(xt.values, expected, real)
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected, expected_value",
-    [
-        (specs.features.valid(DiscreteInput).obj(values=[1, 2, 3]), False, None),
-    ],
-)
-def test_discrete_input_feature_is_fixed(input_feature, expected, expected_value):
-    print(input_feature)
-    assert input_feature.is_fixed() == expected
-    assert input_feature.fixed_value() == expected_value
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected_lower, expected_upper",
-    [
-        (
-            specs.features.valid(DiscreteInput).obj(values=[1.0, 2.0, 3.0]),
-            1,
-            3,
-        ),
-    ],
-)
-def test_discrete_input_feature_bounds(input_feature, expected_lower, expected_upper):
-    assert input_feature.upper_bound == expected_upper
-    assert input_feature.lower_bound == expected_lower
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected",
-    [
-        (
-            DiscreteInput(key="if1", values=[2.0, 3.0]),
-            (1.0, 4.0),
-        ),
-        (
-            DiscreteInput(key="if1", values=[0.0, 3.0]),
-            (0.0, 4.0),
-        ),
-        (
-            DiscreteInput(key="if1", values=[2.0, 5.0]),
-            (1.0, 5.0),
-        ),
-    ],
-)
-def test_discrete_input_feature_get_bounds(input_feature, expected):
-    experiments = pd.DataFrame(
-        {"if1": [1.0, 2.0, 3.0, 4.0], "if2": [1.0, 1.0, 1.0, 1.0]}
-    )
-    lower, upper = input_feature.get_bounds(values=experiments[input_feature.key])
-    assert (lower[0], upper[0]) == expected
-    lower, upper = input_feature.get_bounds()
-    assert (lower[0], upper[0]) == (
-        input_feature.lower_bound,
-        input_feature.upper_bound,
-    )
-
-
-@pytest.mark.parametrize(
-    "input_feature, values",
-    [
-        (
-            specs.features.valid(DiscreteInput).obj(values=[1, 2, 3]),
-            pd.Series([random.choice([1, 2, 3]) for _ in range(20)]),
-        ),
-    ],
-)
-def test_discrete_input_feature_validate_candidental_valid(input_feature, values):
-    input_feature.validate_candidental(values)
-
-
-@pytest.mark.parametrize(
-    "input_feature, values",
-    [
-        (
-            specs.features.valid(DiscreteInput).obj(values=[1, 2]),
-            pd.Series([1, 2, 3]),
-        ),
-    ],
-)
-def test_discrete_input_feature_validate_candidental_invalid(input_feature, values):
-    with pytest.raises(ValueError):
-        input_feature.validate_candidental(values)
-
-
-def test_from_continuous():
-    d = DiscreteInput(key="d", values=[1, 2, 3])
-
-    continuous_values = pd.DataFrame(
-        columns=["d"],
-        data=[1.8, 1.7, 2.9, 1.9],
-    )
-    samples = d.from_continuous(continuous_values)
-    assert np.all(samples == pd.Series([2, 2, 3, 2]))
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected",
-    [
-        (
-            CategoricalInput(key="if1", categories=["a", "b"], allowed=[True, True]),
-            ["a", "b"],
-        ),
-        (
-            CategoricalInput(key="if2", categories=["a", "b"], allowed=[True, True]),
-            ["a", "b"],
-        ),
-        (
-            CategoricalInput(key="if3", categories=["a", "b"], allowed=[True, False]),
-            ["a"],
-        ),
-        (
-            CategoricalInput(key="if4", categories=["a", "b"], allowed=[False, True]),
-            ["b"],
-        ),
-    ],
-)
-def test_categorical_input_feature_get_possible_categories(input_feature, expected):
-    experiments = pd.DataFrame(
-        {"if1": ["a", "b"], "if2": ["a", "a"], "if3": ["a", "a"], "if4": ["b", "b"]}
-    )
-    categories = input_feature.get_possible_categories(experiments[input_feature.key])
-    assert categories == expected
-
-
-@pytest.mark.parametrize(
-    "input_feature, values, strict",
-    [
-        (
-            specs.features.valid(CategoricalInput).obj(categories=["a", "b", "c"]),
-            pd.Series([random.choice(["a", "b", "c"]) for _ in range(20)]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(categories=["a", "b", "c"]),
-            pd.Series([random.choice(["a", "b", "c"]) for _ in range(20)]),
-            False,
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["a", "b", "c"],
-                allowed=[True, False, False],
-            ),
-            pd.Series([random.choice(["a", "b", "c"]) for _ in range(20)]),
-            # CategoricalInput(**VALID_FIXED_CATEGORICAL_INPUT_FEATURE_SPEC),
-            # pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["1", "2", "3"],
-                allowed=[True, False, False],
-            ),
-            pd.Series([random.choice([1, 2, 3]) for _ in range(20)]),
-            # CategoricalInput(**VALID_FIXED_CATEGORICAL_INPUT_FEATURE_SPEC),
-            # pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            False,
-        ),
-    ],
-)
-def test_categorical_input_feature_validate_valid(input_feature, values, strict):
-    input_feature.validate_experimental(values, strict)
-
-
-@pytest.mark.parametrize(
-    "input_feature, values, strict",
-    [
-        (
-            specs.features.valid(CategoricalInput).obj(categories=["a", "b", "c"]),
-            pd.Series(["a", "b", "c", "d"]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(categories=["a", "b", "c"]),
-            pd.Series(["a", "b", "c", "d"]),
-            False,
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["a", "b", "c"],
-                allowed=[True, False, False],
-            ),
-            pd.Series(["a", "a"]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["a", "b", "c"],
-                allowed=[True, False, False],
-            ),
-            pd.Series(["a", "b"]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["1", "2", "3"],
-                allowed=[True, False, False],
-            ),
-            pd.Series([1, 2]),
-            # CategoricalInput(**VALID_FIXED_CATEGORICAL_INPUT_FEATURE_SPEC),
-            # pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["one", "two", "three"],
-                allowed=[True, False, False],
-            ),
-            pd.Series([1, 2, 3]),
-            # CategoricalInput(**VALID_FIXED_CATEGORICAL_INPUT_FEATURE_SPEC),
-            # pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            True,
-        ),
-    ],
-)
-def test_categorical_input_feature_validate_invalid(input_feature, values, strict):
-    with pytest.raises(ValueError):
-        input_feature.validate_experimental(values, strict)
-
-
-@pytest.mark.parametrize(
-    "input_feature, values",
-    [
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["c1", "c2", "c3"],
-                allowed=[True, True, True],
-            ),
-            pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["a", "b"], allowed=[True, False]
-            ),
-            pd.Series(["a", "a"]),
-        ),
-    ],
-)
-def test_categorical_input_feature_validate_candidental_valid(input_feature, values):
-    input_feature.validate_candidental(values)
-
-
-@pytest.mark.parametrize(
-    "input_feature, values",
-    [
-        (
-            specs.features.valid(CategoricalInput).obj(categories=["a", "b", "c"]),
-            pd.Series(["a", "b", "c", "d"]),
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["a", "b"], allowed=[True, False]
-            ),
-            pd.Series(["a", "b"]),
-        ),
-    ],
-)
-def test_categorical_input_feature_validate_candidental_invalid(input_feature, values):
-    with pytest.raises(ValueError):
-        input_feature.validate_candidental(values)
-
-
-@pytest.mark.parametrize(
-    "key, categories, samples",
-    [
-        ("c", ["B", "A", "C"], ["A", "A", "C", "B"]),
-        ("c_alpha", ["B_b", "_A_a", "C_c_"], ["_A_a", "_A_a", "C_c_", "B_b"]),
-        (
-            "__c_alpha_c_",
-            ["__c_alpha_c__B_b", "__c_alpha_c___A_a", "__c_alpha_c__C_c_"],
-            [
-                "__c_alpha_c___A_a",
-                "__c_alpha_c___A_a",
-                "__c_alpha_c__C_c_",
-                "__c_alpha_c__B_b",
-            ],
-        ),
-    ],
-)
-def test_categorical_to_one_hot_encoding(key, categories, samples):
-    c = CategoricalInput(key=key, categories=categories)
-    samples = pd.Series(samples)
-    t_samples = c.to_onehot_encoding(samples)
-    assert_frame_equal(
-        t_samples,
-        pd.DataFrame(
-            data=[[0.0, 1.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, 1.0], [1.0, 0.0, 0.0]],
-            columns=[f"{key}_{cat_str}" for cat_str in categories],
-        ),
-    )
-    untransformed = c.from_onehot_encoding(t_samples)
-    assert np.all(samples == untransformed)
-
-
-@pytest.mark.parametrize(
-    "key, categories",
-    [
-        ("c", ["B", "A", "C"]),
-        ("c_alpha", ["B_b", "_A_a", "C_c_"]),
-        (
-            "__c_alpha_c_",
-            ["__c_alpha_c__B_b", "__c_alpha_c___A_a", "__c_alpha_c__C_c_"],
-        ),
-    ],
-)
-def test_categorical_from_one_hot_encoding(key, categories):
-    c = CategoricalInput(key=key, categories=categories)
-    one_hot_values = pd.DataFrame(
-        columns=[f"{key}_{cat_str}" for cat_str in categories] + ["misc"],
-        data=[[0.9, 0.4, 0.2, 6], [0.8, 0.7, 0.9, 9]],
-    )
-    samples = c.from_onehot_encoding(one_hot_values)
-    assert np.all(samples == pd.Series([categories[0], categories[2]]))
-
-
-def test_categorical_from_one_hot_encoding_invalid():
-    c = CategoricalInput(key="c", categories=["B", "A", "C"])
-    one_hot_values = pd.DataFrame(
-        columns=["c_B", "c_A", "misc"],
-        data=[
-            [
-                0.9,
-                0.4,
-                0.2,
-            ],
-            [0.8, 0.7, 0.9],
-        ],
-    )
-    with pytest.raises(ValueError):
-        c.from_onehot_encoding(one_hot_values)
-
-
-@pytest.mark.parametrize(
-    "key, categories, samples",
-    [
-        ("c", ["B", "A", "C"], ["A", "A", "C", "B"]),
-        ("c_alpha", ["B_b", "_A_a", "C_c_"], ["_A_a", "_A_a", "C_c_", "B_b"]),
-        (
-            "__c_alpha_c_",
-            ["__c_alpha_c__B_b", "__c_alpha_c___A_a", "__c_alpha_c__C_c_"],
-            [
-                "__c_alpha_c___A_a",
-                "__c_alpha_c___A_a",
-                "__c_alpha_c__C_c_",
-                "__c_alpha_c__B_b",
-            ],
-        ),
-    ],
-)
-def test_categorical_to_dummy_encoding(key, categories, samples):
-    c = CategoricalInput(key=key, categories=categories)
-    samples = pd.Series(samples)
-    t_samples = c.to_dummy_encoding(samples)
-    assert_frame_equal(
-        t_samples,
-        pd.DataFrame(
-            data=[[1.0, 0.0], [1.0, 0.0], [0.0, 1.0], [0.0, 0.0]],
-            columns=[f"{key}_{cat_str}" for cat_str in categories[1:]],
-        ),
-    )
-    untransformed = c.from_dummy_encoding(t_samples)
-    assert np.all(samples == untransformed)
-
-
-@pytest.mark.parametrize(
-    "key, categories",
-    [
-        ("c", ["B", "A", "C"]),
-        ("c_alpha", ["B_b", "_A_a", "C_c_"]),
-        (
-            "__c_alpha_c_",
-            ["__c_alpha_c__B_b", "__c_alpha_c___A_a", "__c_alpha_c__C_c_"],
-        ),
-    ],
-)
-def test_categorical_from_dummy_encoding(key, categories):
-    c = CategoricalInput(key=key, categories=categories)
-    one_hot_values = pd.DataFrame(
-        columns=[f"{key}_{cat_str}" for cat_str in categories[1:]] + ["misc"],
-        data=[[0.9, 0.05, 6], [0.1, 0.1, 9]],
-    )
-    samples = c.from_dummy_encoding(one_hot_values)
-    assert np.all(samples == pd.Series([categories[1], categories[0]]))
-
-
-def test_categorical_to_label_encoding():
-    c = CategoricalInput(key="c", categories=["B", "A", "C"])
-    samples = pd.Series(["A", "A", "C", "B"])
-    t_samples = c.to_ordinal_encoding(samples)
-    assert_series_equal(t_samples, pd.Series([1, 1, 2, 0], name="c"))
-    untransformed = c.from_ordinal_encoding(t_samples)
-    assert np.all(samples == untransformed)
-
-
-@pytest.mark.parametrize(
-    "feature, transform_type, values, expected",
-    [
-        (
-            CategoricalInput(key="c", categories=["B", "A", "C"]),
-            CategoricalEncodingEnum.ORDINAL,
-            None,
-            (0, 2),
-        ),
-        (
-            CategoricalInput(key="c", categories=["B", "A", "C"]),
-            CategoricalEncodingEnum.ONE_HOT,
-            None,
-            ([0, 0, 0], [1, 1, 1]),
-        ),
-        (
-            CategoricalInput(
-                key="c", categories=["B", "A", "C"], allowed=[True, False, True]
-            ),
-            CategoricalEncodingEnum.ONE_HOT,
-            pd.Series(["A", "B", "C"]),
-            ([0, 0, 0], [1, 1, 1]),
-        ),
-        (
-            CategoricalInput(
-                key="c", categories=["B", "A", "C"], allowed=[True, False, True]
-            ),
-            CategoricalEncodingEnum.ONE_HOT,
-            None,
-            ([0, 0, 0], [1, 0, 1]),
-        ),
-        (
-            CategoricalInput(key="c", categories=["B", "A", "C"]),
-            CategoricalEncodingEnum.DUMMY,
-            None,
-            ([0, 0], [1, 1]),
-        ),
-    ],
-)
-def test_categorical_get_bounds(feature, transform_type, values, expected):
-    lower, upper = feature.get_bounds(transform_type=transform_type, values=values)
-    assert np.allclose(lower, expected[0])
-    assert np.allclose(upper, expected[1])
-    # test the same for the categorical with descriptor
-    f = CategoricalDescriptorInput(
-        key="c",
-        categories=feature.categories,
-        allowed=feature.allowed,
-        descriptors=["alpha", "beta"],
-        values=[[1, 2], [3, 4], [5, 6]],
-    )
-    lower, upper = f.get_bounds(transform_type=transform_type, values=values)
-    assert np.allclose(lower, expected[0])
-    assert np.allclose(upper, expected[1])
-
-
-@pytest.mark.parametrize(
-    "key, categories, samples_in, descriptors",
-    [
-        ("c", ["B", "A", "C"], ["A", "A", "C", "B"], ["d1", "d2"]),
-        (
-            "c_alpha",
-            ["B_b", "_A_a", "C_c_"],
-            ["_A_a", "_A_a", "C_c_", "B_b"],
-            ["_d1_d", "d2_d_2_"],
-        ),
-        (
-            "__c_alpha_c_",
-            ["__c_alpha_c__B_b", "__c_alpha_c___A_a", "__c_alpha_c__C_c_"],
-            [
-                "__c_alpha_c___A_a",
-                "__c_alpha_c___A_a",
-                "__c_alpha_c__C_c_",
-                "__c_alpha_c__B_b",
-            ],
-            ["__c_alpha_c__d1_d", "__c_alpha_c_d2_d_2_"],
-        ),
-    ],
-)
-def test_categorical_descriptor_to_descriptor_encoding(
-    key, categories, samples_in, descriptors
-):
-    c = CategoricalDescriptorInput(
-        key=key,
-        categories=categories,
-        descriptors=descriptors,
-        values=[[1, 2], [3, 4], [5, 6]],
-    )
-    samples = pd.Series(samples_in)
-    t_samples = c.to_descriptor_encoding(samples)
-    assert_frame_equal(
-        t_samples,
-        pd.DataFrame(
-            data=[[3.0, 4.0], [3.0, 4.0], [5.0, 6.0], [1.0, 2.0]],
-            columns=[f"{key}_{des_str}" for des_str in descriptors],
-        ),
-    )
-    untransformed = c.from_descriptor_encoding(t_samples)
-    assert np.all(samples == untransformed)
-
-
-@pytest.mark.parametrize(
-    "key, categories, descriptors",
-    [
-        ("c", ["B", "A", "C"], ["d1", "d2"]),
-        ("c_alpha", ["B_b", "_A_a", "C_c_"], ["_d1_d", "d2_d_2_"]),
-        (
-            "__c_alpha_c_",
-            ["__c_alpha_c__B_b", "__c_alpha_c___A_a", "__c_alpha_c__C_c_"],
-            ["__c_alpha_c__d1_d", "__c_alpha_c_d2_d_2_"],
-        ),
-    ],
-)
-def test_categorical_descriptor_from_descriptor_encoding(key, categories, descriptors):
-    c1 = CategoricalDescriptorInput(
-        key=key,
-        categories=categories,
-        descriptors=descriptors,
-        values=[[1, 2], [3, 4], [5, 6]],
-    )
-    descriptor_values = pd.DataFrame(
-        columns=[f"{key}_{des_str}" for des_str in descriptors] + ["misc"],
-        data=[[1.05, 2.5, 6], [4, 4.5, 9]],
-    )
-    samples = c1.from_descriptor_encoding(descriptor_values)
-    print(samples)
-    assert np.all(samples == pd.Series([categories[0], categories[1]]))
-
-    c2 = CategoricalDescriptorInput(
-        key=key,
-        categories=categories,
-        descriptors=descriptors,
-        values=[[1, 2], [3, 4], [5, 6]],
-        allowed=[False, True, True],
-    )
-
-    samples = c2.from_descriptor_encoding(descriptor_values)
-    print(samples)
-    assert np.all(samples == pd.Series([categories[1], categories[1]]))
-
-
-def test_categorical_descriptor_to_descriptor_encoding_1d():
-    c = CategoricalDescriptorInput(
-        key="c",
-        categories=["B", "A", "C"],
-        descriptors=["d1"],
-        values=[[1], [3], [5]],
-    )
-    samples = pd.Series(["A", "A", "C", "B"])
-    t_samples = c.to_descriptor_encoding(samples)
-    assert_frame_equal(
-        t_samples,
-        pd.DataFrame(
-            data=[[3.0], [3.0], [5.0], [1.0]],
-            columns=["c_d1"],
-        ),
-    )
-    untransformed = c.from_descriptor_encoding(t_samples)
-    assert np.all(samples == untransformed)
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected_with_values, expected",
-    [
-        (
-            CategoricalDescriptorInput(
-                key="if1",
-                categories=["a", "b"],
-                allowed=[True, True],
-                descriptors=["alpha", "beta"],
-                values=[[1, 2], [3, 4]],
-            ),
-            ([1, 2], [3, 4]),
-            ([1, 2], [3, 4]),
-        ),
-        (
-            CategoricalDescriptorInput(
-                key="if2",
-                categories=["a", "b", "c"],
-                allowed=[True, False, True],
-                descriptors=["alpha", "beta"],
-                values=[[1, 2], [3, 4], [1, 5]],
-            ),
-            ([1, 2], [3, 5]),
-            ([1, 2], [1, 5]),
-        ),
-        # (CategoricalInput(key="if2", categories = ["a","b"], allowed = [True, True]), ["a","b"]),
-        # (CategoricalInput(key="if3", categories = ["a","b"], allowed = [True, False]), ["a"]),
-        # (CategoricalInput(key="if4", categories = ["a","b"], allowed = [True, False]), ["a", "b"]),
-        # (ContinuousInput(key="if1", lower_bound=2.5, upper_bound=2.9), (1,3.)),
-        # (ContinuousInput(key="if2", lower_bound=1., upper_bound=3.), (1,3.)),
-        # (ContinuousInput(key="if2", lower_bound=1., upper_bound=1.), (1,1.)),
-    ],
-)
-def test_categorical_descriptor_feature_get_bounds(
-    input_feature, expected_with_values, expected
-):
-    experiments = pd.DataFrame(
-        {"if1": ["a", "b"], "if2": ["a", "c"], "if3": ["a", "a"], "if4": ["b", "b"]}
-    )
-    lower, upper = input_feature.get_bounds(
-        transform_type=CategoricalEncodingEnum.DESCRIPTOR,
-        values=experiments[input_feature.key],
-    )
-    assert np.allclose(lower, expected_with_values[0])
-    assert np.allclose(upper, expected_with_values[1])
-    lower, upper = input_feature.get_bounds(
-        transform_type=CategoricalEncodingEnum.DESCRIPTOR,
-        values=None,
-    )
-    assert np.allclose(lower, expected[0])
-    assert np.allclose(upper, expected[1])
-
-
-@pytest.mark.parametrize(
-    "input_feature, values, strict",
-    [
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(
-                categories=["c1", "c2", "c3"]
-            ),
-            pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(
-                categories=["c1", "c2", "c3"]
-            ),
-            pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            False,
-        ),
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(
-                categories=["c1", "c2", "c3"],
-                allowed=[True, False, False],
-            ),
-            pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(
-                categories=["c1", "c2", "c3"],
-                allowed=[True, False, False],
-            ),
-            pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            False,
-        ),
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(
-                categories=["c1", "c2", "c3"],
-                allowed=[True, False, False],
-            ),
-            pd.Series(["c1", "c1"]),
-            False,
-        ),
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(
-                categories=["c1", "c2", "c3"],
-                allowed=[False, True, True],
-                descriptors=["d1", "d2"],
-                values=[
-                    [1, 2],
-                    [3, 7],
-                    [3, 1],
-                ],
-            ),
-            pd.Series(["c2", "c3"]),
-            False,
-        ),
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(
-                categories=["1", "2", "3"],
-                allowed=[True, False, False],
-            ),
-            pd.Series([random.choice([1, 2, 3]) for _ in range(20)]),
-            # CategoricalInput(**VALID_FIXED_CATEGORICAL_INPUT_FEATURE_SPEC),
-            # pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            False,
-        ),
-    ],
-)
-def test_categorical_descriptor_input_feature_validate_valid(
-    input_feature, values, strict
-):
-    input_feature.validate_experimental(values, strict)
-
-
-@pytest.mark.parametrize(
-    "input_feature, values, strict",
-    [
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(),
-            pd.Series(["c1", "c4"]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(),
-            pd.Series(["c1", "c4"]),
-            False,
-        ),
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(
-                categories=["c1", "c2", "c3"],
-                allowed=[True, False, False],
-                descriptors=["d1", "d2"],
-                values=[
-                    [1, 2],
-                    [3, 7],
-                    [5, 1],
-                ],
-            ),
-            pd.Series(["c1", "c1"]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalDescriptorInput).obj(
-                categories=["c1", "c2", "c3"],
-                allowed=[False, True, True],
-                descriptors=["d1", "d2"],
-                values=[
-                    [1, 2],
-                    [3, 7],
-                    [3, 1],
-                ],
-            ),
-            pd.Series(["c2", "c3"]),
-            True,
-        ),
-        (
-            specs.features.valid(CategoricalInput).obj(
-                categories=["1", "2", "3"],
-                allowed=[True, False, False],
-            ),
-            pd.Series([1, 2]),
-            # CategoricalInput(**VALID_FIXED_CATEGORICAL_INPUT_FEATURE_SPEC),
-            # pd.Series([random.choice(["c1", "c2", "c3"]) for _ in range(20)]),
-            True,
-        ),
-    ],
-)
-def test_categorical_descriptor_input_feature_validate_invalid(
-    input_feature, values, strict
-):
-    with pytest.raises(ValueError):
-        input_feature.validate_experimental(values, strict)
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected, expected_value, transform_type",
-    [
-        (
-            CategoricalInput(key="k", categories=categories, allowed=allowed),
-            expected,
-            expected_value,
-            transform_type,
-        )
-        for categories, allowed, expected, expected_value, transform_type in [
-            (["1", "2"], None, False, None, None),
-            (["1", "2", "3", "4"], [True, False, False, False], True, ["1"], None),
-            (["1", "2", "3", "4"], [True, True, False, True], False, None, None),
-            (
-                ["1", "2", "3", "4"],
-                [True, False, False, False],
-                True,
-                [0],
-                CategoricalEncodingEnum.ORDINAL,
-            ),
-            (
-                ["1", "2", "3", "4"],
-                [True, False, False, False],
-                True,
-                [1, 0, 0, 0],
-                CategoricalEncodingEnum.ONE_HOT,
-            ),
-            (
-                ["1", "2", "3", "4"],
-                [True, False, False, False],
-                True,
-                [0, 0, 0],
-                CategoricalEncodingEnum.DUMMY,
-            ),
-        ]
-    ]
-    + [
-        (
-            CategoricalDescriptorInput(
-                key="k",
-                categories=["1", "2", "3"],
-                allowed=[True, False, False],
-                descriptors=["alpha", "beta"],
-                values=[[1, 2], [3, 4], [5, 6]],
-            ),
-            expected,
-            expected_value,
-            transform_type,
-        )
-        for expected, expected_value, transform_type in [
-            (True, [1, 2], CategoricalEncodingEnum.DESCRIPTOR)
-        ]
-    ],
-)
-def test_categorical_input_feature_is_fixed(
-    input_feature, expected, expected_value, transform_type
-):
-    assert input_feature.is_fixed() == expected
-    assert input_feature.fixed_value(transform_type) == expected_value
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected",
-    [
-        (
-            CategoricalInput(key="k", categories=categories, allowed=allowed),
-            expected,
-        )
-        for categories, allowed, expected in [
-            (["a", "b", "c"], [True, True, True], ["a", "b", "c"]),
-            (["a", "b", "c"], [False, True, True], ["b", "c"]),
-        ]
-    ],
-)
-def test_categorical_input_feature_allowed_categories(input_feature, expected):
-    assert input_feature.get_allowed_categories() == expected
-
-
-@pytest.mark.parametrize(
-    "input_feature, expected",
-    [
-        (
-            CategoricalInput(key="k", categories=categories, allowed=allowed),
-            expected,
-        )
-        for categories, allowed, expected in [
-            (["a", "b", "c"], [True, True, True], []),
-            (["a", "b", "c"], [False, True, True], ["a"]),
-        ]
-    ],
-)
-def test_categorical_input_feature_forbidden_categories(input_feature, expected):
-    assert input_feature.get_forbidden_categories() == expected
-
-
-@pytest.mark.parametrize(
-    "categories, descriptors, values",
-    [
-        (["c1", "c2"], ["d1", "d2", "d3"], [[1, 2, 3], [4, 5, 6]]),
-        (
-            ["c1", "c2", "c3", "c4"],
-            ["d1", "d2", "d3"],
-            [
-                [1, 2, 3],
-                [4, 5, 6],
-                [4, 5, 6],
-                [4, 5, 6],
-            ],
-        ),
-    ],
-)
-def test_categorical_descriptor_input_feature_as_dataframe(
-    categories, descriptors, values
-):
-    f = CategoricalDescriptorInput(
-        key="k", categories=categories, descriptors=descriptors, values=values
-    )
-    df = f.to_df()
-    assert len(df.columns) == len(descriptors)
-    assert len(df) == len(categories)
-    assert df.values.tolist() == values
-
-
-@pytest.mark.parametrize(
-    "descriptors, values",
-    [
-        (["a", "b"], [1.0, 2.0]),
-        (["a", "b", "c"], [1.0, 2.0, 3.0]),
-    ],
-)
-def test_continuous_descriptor_input_feature_as_dataframe(descriptors, values):
-    f = ContinuousDescriptorInput(
-        key="k",
-        bounds=(1, 2),
-        descriptors=descriptors,
-        values=values,
-    )
-    df = f.to_df()
-    assert len(df.columns) == len(descriptors)
-    assert len(df) == 1
-    assert df.values.tolist()[0] == values
-
-
-@pytest.mark.parametrize(
-    "categories, descriptors, values",
-    [
-        (["c1", "c2"], ["d1", "d2", "d3"], [[1, 2, 3], [4, 5, 6]]),
-        (
-            ["c1", "c2", "c3", "c4"],
-            ["d1", "d2", "d3"],
-            [
-                [1, 2, 3],
-                [4, 5, 6],
-                [4, 5, 6],
-                [4, 5, 6],
-            ],
-        ),
-    ],
-)
-def test_categorical_descriptor_input_feature_from_dataframe(
-    categories, descriptors, values
-):
-    df = pd.DataFrame.from_dict(
-        dict(zip(categories, values)),
-        orient="index",
-        columns=descriptors,
-    )
-    f = CategoricalDescriptorInput.from_df("k", df)
-    assert f.categories == categories
-    assert f.descriptors == descriptors
-    assert f.values == values
-
-
-cont = specs.features.valid(ContinuousInput).obj()
-cat = specs.features.valid(CategoricalInput).obj()
-cat_ = specs.features.valid(CategoricalDescriptorInput).obj()
-out = specs.features.valid(ContinuousOutput).obj()
-
-
-@pytest.mark.parametrize(
-    "unsorted_list, sorted_list",
-    [
-        (
-            [cont, cat_, cat, out],
-            [cont, cat_, cat, out],
-        ),
-        (
-            [cont, cat_, cat, out, cat_, out],
-            [cont, cat_, cat_, cat, out, out],
-        ),
-        (
-            [cont, out],
-            [cont, out],
-        ),
-        (
-            [out, cont],
-            [cont, out],
-        ),
-    ],
-)
-def test_feature_sorting(unsorted_list, sorted_list):
-    assert sorted(unsorted_list) == sorted_list
 
 
 # test features container
